Refactor FieldType class.
